MA-22568-1 dma-buf/heaps: Correct the types of fd_flags and heap_flags
Casting both fd_flags and heap_flags into unsigned long.
The following problems will be encountered:

CFI failure at dma_heap_ioctl+0x210/0x3d8 (target: dsp_heap_allocate+0x0/0x124 [dsp_heap]; expected type: 0xec67cdc9)

+++ drivers/dma-buf/heaps/dsp_heap.c
@@ -81,8 +81,8 @@ static const struct dma_buf_ops dsp_heap_buf_ops = {
 
 static struct dma_buf * dsp_heap_allocate(struct dma_heap *heap,
 					  unsigned long len,
-					  unsigned long fd_flags,
-					  unsigned long heap_flags)
+					  u32 fd_flags,
+					  u64 heap_flags)
 {
 	struct dsp_heap *dsp_heap = dma_heap_get_drvdata(heap);
 	DEFINE_DMA_BUF_EXPORT_INFO(exp_info);
